+++ example/FamilyHubs.Example/Pages/Examples/SummaryList/Index.cshtml
@@ -10,11 +10,16 @@
 @* todo: version that supports action column *@
 
 <dl class="govuk-summary-list">
-    <partial name="_SummaryListRow" model='("Name", "Sarah Philips")' />
-    <partial name="_SummaryListRows" model='("Address", (IEnumerable<string>)new[] {"72 Guild Street","London"})' />
+    <summary-list-row key="Empty (row shouldn't render)"></summary-list-row>
+    <summary-list-row key="Name">Sarah Philips</summary-list-row>
+    @* todo: helper for converting IEnumerable<string> to use <br/>?*@
+    <summary-list-row key="Address">72 Guild Street <br/> London</summary-list-row>
+
+    <partial name="_SummaryListRow" model='("Name", "Sarah Philips")'/>
+    <partial name="_SummaryListRows" model='("Address", (IEnumerable<string>) new[] {"72 Guild Street", "London"})'/>
     @*todo: default not nice*@
-    <partial name="_SummaryListLink" model='("Phone", "07700 900457", LinkType.Phone, default(string?))' />
-    <partial name="_SummaryListLink" model='("Email", "sarah.phillips@example.com", LinkType.Email, default(string?))' />
-    <partial name="_SummaryListLink" model='("Website", "https://www.example.com", LinkType.WebPage, "Example website")' />
-    <partial name="_SummaryListLink" model='("Website", "https://www.example.com", LinkType.WebPageInNewTab, "Example website")' />
+    <partial name="_SummaryListLink" model='("Phone", "07700 900457", LinkType.Phone, default(string?))'/>
+    <partial name="_SummaryListLink" model='("Email", "sarah.phillips@example.com", LinkType.Email, default(string?))'/>
+    <partial name="_SummaryListLink" model='("Website", "https://www.example.com", LinkType.WebPage, "Example website")'/>
+    <partial name="_SummaryListLink" model='("Website", "https://www.example.com", LinkType.WebPageInNewTab, "Example website")'/>
 </dl>
